Do y2019 d16
Had to look up how to do part 2, since

- my brute force was way too slow
- I didn't think to make sure the offset was past the midway point
- I didn't consider what happens when the offset is past the midway
  point

What happens is this: the ith digit is just the sum of all the digits
from i to the end. Since the pattern is all 0s before the ith digit and
all 1s after.

Also, forgetting to make sum a u32 in part 2 got me with overflow bugs
in release mode.